moves styles from ProductEditRow
