+++ src/user/modules/ProductsEditListContainer/ProductEditRow.tsx
@@ -1,66 +1,19 @@
 import React, { useState } from "react";
-import styled from "styled-components";
-import styledTS from "styled-components-ts";
-import { CardRow } from "src/common/ui/Card";
-import { TextTable } from "src/common/ui/Text";
-
-const UnitsContainer = styled.div`
-  display: grid;
-  grid-gap: 0.5rem;
-  grid-template-columns: 1rem 2.5rem 1rem;
-`;
-const Icons = styled.img`
-  width: 1rem;
-  height: 1rem;
-  margin: auto 0px;
-`;
-const InputUnits = styled.input`
-  width: 2.5rem;
-  height: 2rem;
-  box-sizing: border-box;
-  text-align: center;
-`;
-const DivWrapper = styled.div`
-  clip-path: inset(-1px);
-`;
-
-const CustomRow = styled(CardRow)`
-  position: relative;
-  display: grid;
-  grid-template-columns: 1fr 1fr fit-content(5.5rem) 0px;
-  grid-gap: 0.125rem;
-`;
-
-interface ITextTable {
-  textAlign?: string;
-}
-const TextTableProduct = styledTS<ITextTable>(styled(TextTable))`
-  margin: auto 0px;
-  text-align :${({ textAlign }) => textAlign || "center"};
-`;
-
-const ExtraButtonDivRelative = styled.div`
-  position: relative;
-  right: -1rem;
-`;
-const ExtraButtonInnerDivAbsolute = styled.div`
-  position: absolute;
-  height: 3.125rem;
-  top: -0.4875rem;
-  display: flex;
-  background: #ff6200;
-`;
-
-const DivDelete = styled.div`
-  width: 1rem;
-  height: 1rem;
-  background-image: url(${require("src/images/icons/delete_white.svg")});
-  background-size: 1rem;
-  margin: auto 0px auto 1rem;
-`;
+import { IOrderProduct } from "src/user/models/OrderProduct";
+import {
+  UnitsContainer,
+  Icons,
+  InputUnits,
+  DivWrapper,
+  CustomProductRow,
+  TextTableProduct,
+  ExtraButtonDivRelative,
+  ExtraButtonInnerDivAbsolute,
+  DivDelete,
+} from "../ui/ProductEditRow";
 
 interface IProductRowProps {
-  product: any;
+  product: IOrderProduct;
   handleAddUnit: (id: string) => void;
   handleMinusUnit: (id: string) => void;
   handleRemoveProduct: (id: string) => void;
@@ -92,15 +45,20 @@ export const ProductEditRow: React.FC<IProductRowProps> = ({
 
   return (
     <DivWrapper>
-      <CustomRow onTouchMove={onTouchMove} onTouchStart={onTouchStart} style={{ left }}>
+      <CustomProductRow onTouchMove={onTouchMove} onTouchStart={onTouchStart} style={{ left }}>
         <TextTableProduct textAlign="left">{product.name}</TextTableProduct>
         <TextTableProduct textAlign="right">${product.price} MXN</TextTableProduct>
         <UnitsContainer>
           <Icons
             onClick={handleMinusUnit.bind(null, product.id)}
-            src={require(`src/images/icons/remove_circle${product.units == 0 ? "_gray" : ""}.svg`)}
+            src={require(`src/images/icons/remove_circle${
+              product.quantity === 0 ? "_gray" : ""
+            }.svg`)}
           />
-          <InputUnits onChange={handleOnChangeInput.bind(null, product.id)} value={product.units} />
+          <InputUnits
+            onChange={handleOnChangeInput.bind(null, product.id)}
+            value={product.quantity}
+          />
           <Icons
             onClick={handleAddUnit.bind(null, product.id)}
             src={require("src/images/icons/add_circle.svg")}
@@ -114,7 +72,7 @@ export const ProductEditRow: React.FC<IProductRowProps> = ({
             <DivDelete />
           </ExtraButtonInnerDivAbsolute>
         </ExtraButtonDivRelative>
-      </CustomRow>
+      </CustomProductRow>
     </DivWrapper>
   );
 };
