+++ src/user/modules/ui/ProductEditRow/index.ts
@@ -1,0 +1,11 @@
+export {
+  UnitsContainer,
+  Icons,
+  InputUnits,
+  DivWrapper,
+  CustomProductRow,
+  TextTableProduct,
+  ExtraButtonDivRelative,
+  ExtraButtonInnerDivAbsolute,
+  DivDelete,
+} from "./ProductEditRow";
